Relocated the QA and DisplayQuestions components
